PushCatchupStatus: reject duplicate registration, split read/write API
register() used to put-replace a second registration under the same id,
which the annotation layer's own duplicate check made unreachable within
one stack but not across a dual-stack context sharing one PushCatchupStatus
bean. It now throws, matching the fail-fast posture of the annotation
layer's DuplicateSubscriptionIdException family (which this module cannot
depend on directly, common has no subscription dependency by design).

PushCatchupStatus is now a read-only interface (of, isCaughtUp, all).
The mutable methods moved to PushCatchupStatusImpl, which the three
registrar write sites look up by concrete type. The bean is declared by
that concrete type in both starters, the same precedent
SagaInstancesRegistryImpl already established for by-type lookup before
singleton instantiation. Applications injecting PushCatchupStatus by name
are unaffected, only the writer moved.

Fixes #620
