+++ framework/spring-boot-autoconfigure/blocking/src/main/java/org/occurrent/springboot/blocking/OccurrentBlockingAnnotationConfiguration.java
@@ -20,6 +20,7 @@ import org.occurrent.dsl.saga.SagaInstances;
 import org.occurrent.dsl.saga.SagaInstancesRegistry;
 import org.occurrent.dsl.saga.internal.SagaInstancesRegistryImpl;
 import org.occurrent.springboot.common.PushCatchupStatus;
+import org.occurrent.springboot.common.PushCatchupStatusImpl;
 import org.occurrent.springboot.common.OnSubscriptionsNotDisabledCondition;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Bean;
@@ -65,8 +66,12 @@ public class OccurrentBlockingAnnotationConfiguration {
     @Bean
     @ConditionalOnMissingBean(PushCatchupStatus.class)
     @Conditional(OnSubscriptionsNotDisabledCondition.class)
-    public PushCatchupStatus occurrentPushCatchupStatus() {
-        return new PushCatchupStatus();
+    public PushCatchupStatusImpl occurrentPushCatchupStatus() {
+        // The declared return type is the implementation, not the PushCatchupStatus interface an application injects,
+        // so that the registrars' by-type lookup of the writable type matches from the bean definition rather than
+        // only once the singleton has been instantiated, the same reason SagaInstancesRegistryImpl is declared by its
+        // concrete type below.
+        return new PushCatchupStatusImpl();
     }
 
     /**
