+++ framework/spring-boot-autoconfigure/blocking/src/test/java/org/occurrent/springboot/blocking/ProjectionAnnotationValidationTest.java
@@ -31,6 +31,7 @@ import org.occurrent.eventstore.api.blocking.PositionOrderedReader;
 import org.occurrent.eventstore.api.dcb.DcbCriteria;
 import org.occurrent.filter.Filter;
 import org.occurrent.springboot.common.PushCatchupStatus;
+import org.occurrent.springboot.common.PushCatchupStatusImpl;
 import org.occurrent.springboot.common.OccurrentProperties;
 import org.occurrent.subscription.api.blocking.CheckpointStorage;
 import org.occurrent.subscription.push.blocking.PushSubscriptionModel;
@@ -314,8 +315,8 @@ class ProjectionAnnotationValidationTest {
         }
 
         @Bean
-        PushCatchupStatus pushCatchupStatus() {
-            return new PushCatchupStatus();
+        PushCatchupStatusImpl pushCatchupStatus() {
+            return new PushCatchupStatusImpl();
         }
 
         @Bean
