+++ framework/spring-boot-autoconfigure/blocking/src/test/java/org/occurrent/springboot/blocking/PushCatchupReadinessTest.java
@@ -32,6 +32,7 @@ import org.occurrent.eventstore.api.blocking.PositionOrderedReader;
 import org.occurrent.filter.Filter;
 import org.occurrent.springboot.common.OccurrentProperties;
 import org.occurrent.springboot.common.PushCatchupStatus;
+import org.occurrent.springboot.common.PushCatchupStatusImpl;
 import org.occurrent.subscription.api.blocking.CheckpointStorage;
 import org.occurrent.subscription.push.blocking.PushSubscriptionModel;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
@@ -172,8 +173,8 @@ class PushCatchupReadinessTest {
     static class PushConfigurationSupport {
 
         @Bean
-        PushCatchupStatus pushCatchupStatus() {
-            return new PushCatchupStatus();
+        PushCatchupStatusImpl pushCatchupStatus() {
+            return new PushCatchupStatusImpl();
         }
 
         @Bean
